refactor!: move error type to a separate module
