+++ src/linux/launchers/bottles.rs
@@ -344,7 +344,7 @@ mod tests {
     use test_case::test_case;
 
     use super::*;
-    use crate::{data::GamesParsingError, linux::test_utils::get_mock_file_system_path};
+    use crate::{error::GamesParsingError, linux::test_utils::get_mock_file_system_path};
 
     #[test_case(false, ".local/share"; "standard")]
     #[test_case(true, "invalid/data/path"; "flatpak")]
