+++ src/linux/launchers/heroic/amazon.rs
@@ -113,7 +113,7 @@ mod tests {
     use test_case::test_case;
 
     use super::*;
-    use crate::{data::GamesParsingError, linux::test_utils::get_mock_file_system_path};
+    use crate::{error::GamesParsingError, linux::test_utils::get_mock_file_system_path};
 
     #[test_case(false, ".config"; "standard")]
     #[test_case(true, "invalid/data/path"; "flatpak")]
