+++ src/linux/launchers/lutris.rs
@@ -391,7 +391,7 @@ mod tests {
     use test_case::test_case;
 
     use super::*;
-    use crate::{data::GamesParsingError, linux::test_utils::get_mock_file_system_path};
+    use crate::{error::GamesParsingError, linux::test_utils::get_mock_file_system_path};
 
     #[test_case(false, ".config", ".cache"; "standard")]
     #[test_case(false, "invalid/path", ".cache"; "fallback")]
