+++ src/linux/launchers/steam/steam_shortcuts.rs
@@ -15,7 +15,8 @@ use tracing::{error, trace, warn};
 
 use super::{get_steam_dir, get_steam_flatpak_dir, get_steam_launch_command};
 use crate::{
-    data::{Game, GamesParsingError, GamesResult, Launcher, SupportedLaunchers},
+    data::{Game, GamesResult, Launcher, SupportedLaunchers},
+    error::GamesParsingError,
     macros::logs::{debug_fallback_flatpak, debug_path, warn_no_games},
     parsers::{parse_between_double_quotes, parse_not_double_quote},
     utils::{clean_game_title, get_existing_image_path},
